refactor install log format support into hooks/dashboards.py
